pyproject.toml
[project]
name = "cook-inventory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
